Improve compatibility with short high-bitrate signals
Because RSSI is polled and the receive interrupt isn't "armed" until after RSSI exceeds the minimum,
preambles from short signals are oftne missed because they come in before the next RSSI polling interval.
This makes some protocols such as neptune_r900 unable to work unless you get extremely lucky with timing,
and given that it's bad enough that we can only detect messages on 3 out of 50 channels (It appears adjacent channels
to the one tuned will get detected), we need to improve reliability.

Instead of ignoring pulses when RSSI is low, put them in a "pending noise" bucket.
If RSSI is low in the next polling interval, empty that bucket into the main noise bucket, and reset the number of pulses to 0
If RSSI goes high, continue as normal, except we now have the preamble in there.